Remove Library Utils File
