+++ tests/components/test_parser.cpp
@@ -11,9 +11,9 @@
 #include <masm/assembler/parser.h>
 #include <masm/exceptions.h>
 #include <masm/io/fileio.h>
-#include <masm/utils.h>
 
 #include "libmasm/src/assembler/directive.h"
+#include "shared/load_layout.h"
 #include "tests/testing_utilities.h"
