+++ tools/fixture-gen/CMakeLists.txt
@@ -1,4 +1,7 @@
 add_executable(masm-fg fixture_gen.cpp)
+target_include_directories(masm-fg SYSTEM PRIVATE
+        ${CMAKE_SOURCE_DIR}/shared
+)
 target_link_libraries(masm-fg PRIVATE libmasm CLI11::CLI11)
 set_target_properties(masm-fg PROPERTIES
         RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/bin"
